User Public API (#50740)
- Replace bcrypt with pbkdf2 hashing algorithm for local user passwords
- Expose User resource
- Add new imperative API resources to replace norman actions: PasswordChangeRequest, SelfUser and UserRefreshRequest.

+++ pkg/auth/cleanup/service.go
@@ -8,6 +8,7 @@ import (
 
 	v3 "github.com/rancher/rancher/pkg/apis/management.cattle.io/v3"
 	"github.com/rancher/rancher/pkg/auth/api/secrets"
+	"github.com/rancher/rancher/pkg/auth/providers/local/pbkdf2"
 	controllers "github.com/rancher/rancher/pkg/generated/controllers/management.cattle.io/v3"
 	wcorev1 "github.com/rancher/wrangler/v3/pkg/generated/controllers/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -20,6 +21,7 @@ var errAuthConfigNil = errors.New("cannot get auth provider if its config is nil
 // Service performs cleanup of resources associated with an auth provider.
 type Service struct {
 	secretsInterface wcorev1.SecretController
+	secretsCache     wcorev1.SecretCache
 
 	userClient controllers.UserClient
 
@@ -40,6 +42,7 @@ type Service struct {
 func NewCleanupService(secretsInterface wcorev1.SecretController, c controllers.Interface) *Service {
 	return &Service{
 		secretsInterface: secretsInterface,
+		secretsCache:     secretsInterface.Cache(),
 
 		userClient: c.User(),
 
@@ -173,7 +176,11 @@ func (s *Service) deleteUsers(config *v3.AuthConfig) error {
 		providerName := getProviderNameFromPrincipalNames(u.PrincipalIDs...)
 		if providerName == config.Name {
 			// A fully external user (who was never local) has no password.
-			if u.Password == "" {
+			_, err := s.secretsCache.Get(pbkdf2.LocalUserPasswordsNamespace, u.Name)
+			if err != nil && !apierrors.IsNotFound(err) {
+				return fmt.Errorf("failed to get user secret: %w", err)
+			}
+			if u.Password == "" && apierrors.IsNotFound(err) {
 				err := s.userClient.Delete(u.Name, &metav1.DeleteOptions{})
 				if err != nil && !apierrors.IsNotFound(err) {
 					return err
